fix lidar and radar after merge
