+++ carla_ros_bridge/src/carla_ros_bridge/lidar.py
@@ -19,19 +19,9 @@ import struct
 import numpy
 
 from sensor_msgs.msg import PointCloud2, PointField  # pylint: disable=import-error
-import carla_common.transforms as trans
-from carla_ros_bridge.sensor import Sensor
+from carla_ros_bridge.sensor import Sensor, create_cloud
 
 from ros_compatibility import quaternion_from_euler, euler_from_quaternion
-
-ROS_VERSION = int(os.environ.get('ROS_VERSION', 0))
-
-if ROS_VERSION == 1:
-    # pylint: disable=import-error,ungrouped-imports
-    from sensor_msgs.point_cloud2 import create_cloud
-
-_DATATYPES = {}
-_DATATYPES[PointField.FLOAT32] = ('f', 4)
 
 
 class Lidar(Sensor):
@@ -85,34 +75,7 @@ class Lidar(Sensor):
             PointField(name='intensity', offset=12, datatype=PointField.FLOAT32, count=1)
         ]
 
-        if ROS_VERSION == 1:
-            point_cloud_msg = create_cloud(header, fields, lidar_data)
-
-        # -- taken from
-        # http://docs.ros.org/indigo/api/sensor_msgs/html/point__cloud2_8py_source.html
-        elif ROS_VERSION == 2:
-            cloud_struct = struct.Struct(_get_struct_fmt(False, fields))
-
-            buff = ctypes.create_string_buffer(cloud_struct.size * len(lidar_data))
-
-            point_step, pack_into = cloud_struct.size, cloud_struct.pack_into
-            offset = 0
-            for p in lidar_data:
-                pack_into(buff, offset, *p)
-                offset += point_step
-
-            point_cloud_msg = PointCloud2(header=header,
-                                          height=1,
-                                          width=len(lidar_data),
-                                          is_dense=False,
-                                          is_bigendian=False,
-                                          fields=fields,
-                                          point_step=cloud_struct.size,
-                                          row_step=cloud_struct.size * len(lidar_data),
-                                          data=buff.raw)
-
-        # --
-
+        point_cloud_msg = create_cloud(header, fields, lidar_data)
         self.lidar_publisher.publish(point_cloud_msg)
 
 
@@ -154,15 +117,15 @@ class SemanticLidar(Sensor):
         """
         header = self.get_msg_header()
         fields = [
-            PointField('x', 0, PointField.FLOAT32, 1),
-            PointField('y', 4, PointField.FLOAT32, 1),
-            PointField('z', 8, PointField.FLOAT32, 1),
-            PointField('CosAngle', 12, PointField.FLOAT32, 1),
-            PointField('ObjIdx', 16, PointField.UINT32, 1),
-            PointField('ObjTag', 20, PointField.UINT32, 1),
+            PointField(name='x', offset=0, datatype=PointField.FLOAT32, count=1),
+            PointField(name='y', offset=4, datatype=PointField.FLOAT32, count=1),
+            PointField(name='z', offset=8, datatype=PointField.FLOAT32, count=1),
+            PointField(name='CosAngle', offset=12, datatype=PointField.FLOAT32, count=1),
+            PointField(name='ObjIdx', offset=16, datatype=PointField.UINT32, count=1),
+            PointField(name='ObjTag', offset=20, datatype=PointField.UINT32, count=1)
         ]
 
-        lidar_data = numpy.fromstring(carla_lidar_measurement.raw_data,
+        lidar_data = numpy.fromstring(bytes(carla_lidar_measurement.raw_data),
                                       dtype=numpy.dtype([
                                           ('x', numpy.float32),
                                           ('y', numpy.float32),
@@ -177,24 +140,3 @@ class SemanticLidar(Sensor):
         lidar_data['y'] *= -1
         point_cloud_msg = create_cloud(header, fields, lidar_data.tolist())
         self.semantic_lidar_publisher.publish(point_cloud_msg)
-
-# http://docs.ros.org/indigo/api/sensor_msgs/html/point__cloud2_8py_source.html
-
-
-def _get_struct_fmt(is_bigendian, fields, field_names=None):
-    fmt = '>' if is_bigendian else '<'
-
-    offset = 0
-    for field in (f for f in sorted(fields, key=lambda f: f.offset)
-                  if field_names is None or f.name in field_names):
-        if offset < field.offset:
-            fmt += 'x' * (field.offset - offset)
-            offset = field.offset
-        if field.datatype not in _DATATYPES:
-            print('Skipping unknown PointField datatype [%d]' % field.datatype, file=sys.stderr)
-        else:
-            datatype_fmt, datatype_length = _DATATYPES[field.datatype]
-            fmt += field.count * datatype_fmt
-            offset += field.count * datatype_length
-
-    return fmt
