+++ carla_ros_bridge/src/carla_ros_bridge/radar.py
@@ -12,18 +12,9 @@ Classes to handle Carla Radar
 
 from sensor_msgs.msg import PointCloud2, PointField
 
-import sys
-import ctypes
-import os
-import struct
 import numpy as np
 
-# from sensor_msgs.point_cloud2 import create_cloud
-
-_DATATYPES = {}
-_DATATYPES[PointField.FLOAT32] = ('f', 4)
-
-from carla_ros_bridge.sensor import Sensor
+from carla_ros_bridge.sensor import Sensor, create_cloud
 
 
 class Radar(Sensor):
@@ -60,13 +51,14 @@ class Radar(Sensor):
         :param carla_radar_measurement: carla Radar measurement object
         :type carla_radar_measurement: carla.RadarMeasurement
         """
-        fields = [PointField('x', 0, PointField.FLOAT32, 1),
-                  PointField('y', 4, PointField.FLOAT32, 1),
-                  PointField('z', 8, PointField.FLOAT32, 1),
-                  PointField('Range', 12, PointField.FLOAT32, 1),
-                  PointField('Velocity', 16, PointField.FLOAT32, 1),
-                  PointField('AzimuthAngle', 20, PointField.FLOAT32, 1),
-                  PointField('ElevationAngle', 28, PointField.FLOAT32, 1)]
+        fields = [
+            PointField(name='x', offset=0, datatype=PointField.FLOAT32, count=1),
+            PointField(name='y', offset=4, datatype=PointField.FLOAT32, count=1),
+            PointField(name='z', offset=8, datatype=PointField.FLOAT32, count=1),
+            PointField(name='Range', offset=12, datatype=PointField.FLOAT32, count=1),
+            PointField(name='Velocity', offset=16, datatype=PointField.FLOAT32, count=1),
+            PointField(name='AzimuthAngle', offset=20, datatype=PointField.FLOAT32, count=1),
+            PointField(name='ElevationAngle', offset=28, datatype=PointField.FLOAT32, count=1)]
         points = []
         for detection in carla_radar_measurement:
             points.append([detection.depth * np.cos(-detection.azimuth) * np.cos(detection.altitude),
@@ -77,61 +69,3 @@ class Radar(Sensor):
         radar_msg = create_cloud(self.get_msg_header(
             timestamp=carla_radar_measurement.timestamp), fields, points)
         self.radar_publisher.publish(radar_msg)
-
-
-# http://docs.ros.org/indigo/api/sensor_msgs/html/point__cloud2_8py_source.html
-
-def _get_struct_fmt(is_bigendian, fields, field_names=None):
-    fmt = '>' if is_bigendian else '<'
-
-    offset = 0
-    for field in (f for f in sorted(fields, key=lambda f: f.offset)
-                  if field_names is None or f.name in field_names):
-        if offset < field.offset:
-            fmt += 'x' * (field.offset - offset)
-            offset = field.offset
-        if field.datatype not in _DATATYPES:
-            print('Skipping unknown PointField datatype [%d]' % field.datatype)
-        else:
-            datatype_fmt, datatype_length = _DATATYPES[field.datatype]
-            fmt += field.count * datatype_fmt
-            offset += field.count * datatype_length
-
-    return fmt
-
-# https://github.com/ros/common_msgs/blob/noetic-devel/sensor_msgs/src/sensor_msgs/point_cloud2.p
-
-def create_cloud(header, fields, points):
-    """
-    Create a L{sensor_msgs.msg.PointCloud2} message.
-    @param header: The point cloud header.
-    @type  header: L{std_msgs.msg.Header}
-    @param fields: The point cloud fields.
-    @type  fields: iterable of L{sensor_msgs.msg.PointField}
-    @param points: The point cloud points.
-    @type  points: list of iterables, i.e. one iterable for each point, with the
-                   elements of each iterable being the values of the fields for 
-                   that point (in the same order as the fields parameter)
-    @return: The point cloud.
-    @rtype:  L{sensor_msgs.msg.PointCloud2}
-    """
-
-    cloud_struct = struct.Struct(_get_struct_fmt(False, fields))
-
-    buff = ctypes.create_string_buffer(cloud_struct.size * len(points))
-
-    point_step, pack_into = cloud_struct.size, cloud_struct.pack_into
-    offset = 0
-    for p in points:
-        pack_into(buff, offset, *p)
-        offset += point_step
-
-    return PointCloud2(header=header,
-                       height=1,
-                       width=len(points),
-                       is_dense=False,
-                       is_bigendian=False,
-                       fields=fields,
-                       point_step=cloud_struct.size,
-                       row_step=cloud_struct.size * len(points),
-                       data=buff.raw)
